refactor: move Aim and Route into aim_core, make aim_server a dart:io adapter
handle() is now the single request path. The core no longer prints;
the VM adapter logs unhandled errors via onUnhandledError.
