+++ packages/aim_core/lib/aim_core.dart
@@ -5,9 +5,11 @@
 /// WebAssembly. Use `aim_server` to run on the Dart VM with `HttpServer`.
 library;
 
+export 'src/app.dart';
 export 'src/body.dart';
 export 'src/context.dart';
 export 'src/env.dart';
 export 'src/message.dart';
 export 'src/request.dart';
 export 'src/response.dart';
+export 'src/route.dart';
