+++ packages/aim_core/test/unit/headers_test.dart
@@ -1,4 +1,4 @@
-import 'package:aim_server/aim_server.dart';
+import 'package:aim_core/aim_core.dart';
 import 'package:test/test.dart';
 
 void main() {
